pyproject.toml
[project]
name = "opengis-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-gis>=1.0",
    "psycopg2>=2.9.10",
]
